scraping/pyproject.toml
[project]
name = "scraping"
version = "0.1.0"
description = "Web scraping module for sports data collection"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "aiohttp>=3.9.3",
    "asyncio>=3.4.3",
    "requests==2.31.0",
    "selenium==4.16.0",
    "webdriver-manager==4.0.1",
    "sqlalchemy>=2.0.29",
    "pytest>=8.1.1",
    "pandas>=2.2.2",
    "scrapy>=2.11.1",
    "scrapy-playwright>=0.0.34",
    "anthropic>=0.39.0",
    "python-dotenv>=1.0.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["."]
only-include = ["*.py"]

[tool.ruff]
line-length = 120
target-version = "py310"
dummy-variable-rgx = "^(_+|(_+[a-zA-Z0-9_]*[a-zA-Z0-9]+?))$"
select = ["A", "B", "C", "E", "F", "W", "I"]
ignore = [
    "B904",
    "B008",
    "F841",
    "F401",
    "E302",
    "E305",
    "F403",
    "F405",
    "I001",
    "F811",
]

exclude = [
    ".bzr",
    ".direnv",
    ".eggs",
    ".git",
    ".hg",
    ".mypy_cache",
    ".nox",
    ".pants.d",
    ".ruff_cache",
    ".svn",
    ".tox",
    ".venv",
    "__pypackages__",
    "_build",
    "buck-out",
    "build",
    "dist",
    "node_modules",
    "venv",
]

[tool.ruff.pydocstyle]
convention = "google"

[tool.ruff.mccabe]
max-complexity = 10

[tool.black]
line-length = 99
target-version = ['py310']
preview = true
